pyproject.toml
[project]
name = "badminton-academy-chatbot"
version = "0.1.0"
description = "Badminton Academy Chatbot designed to assist students, parents, and visitors by providing instant information about academy programs, schedules, fees, coaches, and registration."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "openai>=2.24.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
